Task1/requirements.txt
streamlit
pandas
google-genai
python-dotenv
scikit-learn